requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro-datasets>=3.0
kedro-viz>=6.7.0
kedro[jupyter]~=0.19.13
notebook
scikit-learn~=1.5.1
setuptools; python_version >= "3.12"
fastapi
uvicorn
scikit-learn
joblib
google-cloud-storage
numpy
pydantic

